converted netSdkHelper class into methods only
